fix(tests): add integration test suite
