+++ package.json
@@ -75,6 +75,7 @@
     "test:unit": "NODE_OPTIONS=\"--import=./register.mjs\" node --test src/**/**.test.ts",
     "test:link": "yarn build && yarn link",
     "test:e2e": "yarn test:cjs && yarn test:esm && yarn test:web",
+    "test:integration": "yarn build:esm && yarn link && cd ./tests/integration && yarn && yarn test",
     "prepare": "husky install",
     "docs:update": "markdown-toc-gen insert README.md",
     "example:esm": "cd examples/esm && yarn && node index.mjs",
@@ -97,6 +98,7 @@
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^6.4.0",
     "arconnect": "^1.0.3",
+    "cross-env": "^7.0.3",
     "dotenv": "^16.4.5",
     "dotenv-cli": "^7.4.2",
     "esbuild": "^0.19.2",
