+++ tests/integration/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "module": "ESNext"
+  },
+  "include": ["index.test.ts"]
+}
