+++ tsconfig.json
@@ -25,5 +25,5 @@
     "strictNullChecks": true
   },
   "include": ["src"],
-  "exclude": ["lib", "node_modules", "bundles", "tests"]
+  "exclude": ["lib", "node_modules", "bundles", "tests", "tools"]
 }
